added tests for experience tracking module refactored some things in experience tracking changed the paging behaviour for leader board and added check for negative numbers fixed rank not being correct for further pages added test-common module to have some common code für tests fixed command creation

+++ abstracto-application/abstracto-modules/experience-tracking/experience-tracking-impl/src/main/java/dev/sheldan/abstracto/experience/config/ExperienceConfig.java
@@ -23,7 +23,7 @@ public class ExperienceConfig {
     /**
      * The default multiplier from the properties file. This is used, when the bot joins a new guild.
      */
-    private Integer expMultiplier;
+    private Double expMultiplier;
 
     /**
      * The defaul maxLevel from the properties file. This configuration applies globally, as the amount of levels does not depend on the server.
